backend/requirements.txt
asgiref==3.7.2
coloredlogs==15.0.1
dj-rest-auth==5.0.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-tenants==3.5.0
djangorestframework==3.14.0
flatbuffers==23.5.26
gunicorn==21.2.0
humanfriendly==10.0
Markdown==3.5.1
mpmath==1.3.0
onnxruntime==1.16.1

packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.8.0
onnxruntime==1.16.1
opencv-python-headless==4.8.1.78
exifread==3.0.0